Little edits for style consistency.
Great doc, I just tweaked a few words for style consistency.
FYI, custom app is the right term! 
The ellipses at lines 479 and 482 are going to show up as red in the output, but that's okay. I'm assuming we left out portions because they are too long and would make the example unclear, so having the ellipses highlighted is okay.