Suggested edits.
Please check line 1922 to make sure I didn't change the meaning.
Is there any way we can make line 3284 as easy to understand? 

Enroll Duo Factor: Enrolls a user with a Duo factor.  The enrollment process starts with an enrollment request to Okta, then continues with a Duo widget or script that handles the enrollment. At the end of the process, Duo makes a callback to Okta to complete the enrollment.

Verify Duo Factor:  Verification of the Duo factor is implemented as an integration between the OKTA API and Duo widget. The request starts with an API call and then callback from the Duo widget and call to the verification endpoint.